refactor: 코드 정리
코드 중에서 안쓰는 코드 정리.
