+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App.tsx';
 import './index.css';
-import { worker } from './mocks/browser.ts';
+// import { worker } from './mocks/browser.ts';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 function renderApp() {
   const root = ReactDOM.createRoot(
